requirements.txt
requests
python-dotenv
pandas
scikit-learn
mlflow
fastapi
uvicorn